Update DemoOfWaypointsWithDirections.py
Start of Integration with User tracking. In the main loop for running I added an if statement to run a function. If the user is out of frame the function will return True, stopping the rover. If the function returns false then the user is in the frame and the rover will continue. The function has yet to be added to the code. I am adding the base work for implementation.